retireval/web_search.py
"""Web search module for retrieving additional evidence from the internet."""

import logging
import requests
from typing import List, Dict, Optional
from urllib.parse import quote

logger = logging.getLogger(__name__)


class WebSearcher:
    """Search the web for additional evidence."""
    
    def __init__(self, api_key: Optional[str] = None):
        """
        Initialize web searcher.
        
        Args:
            api_key: Optional API key for search service
        """
        self.api_key = api_key
        logger.info("WebSearcher initialized")
    
    def search(
        self,
        query: str,
        num_results: int = 5,
        timeout: int = 10
    ) -> List[Dict]:
        """
        Search the web for relevant documents.
        
        Args:
            query: Search query
            num_results: Number of results to return
            timeout: Request timeout in seconds
            
        Returns:
            List of search results with 'title', 'url', 'snippet' keys
        """
        logger.info(f"Searching web for: {query}")
        
        # This is a placeholder implementation
        # In production, use services like:
        # - Google Custom Search API
        # - Bing Search API
        # - SerpAPI
        # - DuckDuckGo Search
        
        logger.warning("Web search not configured. Please implement with actual API.")
        
        return []
    
    def search_and_summarize(
        self,
        query: str,
        num_results: int = 3
    ) -> List[Dict]:
        """
        Search the web and summarize results.
        
        Args:
            query: Search query
            num_results: Number of results to summarize
            
        Returns:
            List of results with summaries
        """
        results = self.search(query, num_results)
        
        # In production, add summarization logic here
        
        return results


if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)
    
    searcher = WebSearcher()
    results = searcher.search("What is photosynthesis?")
    
    for result in results:
        print(f"Title: {result['title']}")
        print(f"URL: {result['url']}")
        print(f"Snippet: {result['snippet']}\n")
