requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
spacy
wikipedia
pdfplumber
gradio
numpy
pandas
pyyaml
